Update readme.md
Hi @Dalek-Sec @dhanus, here is a PR for an Eng Onboarding readme doc. Formatting isn't where it should to be and there are some more lines I'd like to add. However, it'd be helpful for you to glance over the content and see if it is clear. If you see any major issues, please feel free to let me know!